refactor: rename package "regular" to "embedded"
In our documentation we talk about the embedded engine. It makes sense to name our packages in a similar way.
